requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-bootstrap-form==3.4
django-filter==2.0.0
gunicorn==19.9.0
pytz==2018.5
psycopg2==2.7.5
whitenoise==4.1
